# REFACTOR-#3519: align `mask` interface with its docstring