--- a/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition_manager.py
+++ b/modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition_manager.py
@@ -395,6 +395,8 @@ class PandasOnRayFramePartitionManager(GenericRayFramePartitionManager):
         row_partitions_list,
         col_partitions_list,
         item_to_distribute=None,
+        row_lengths=None,
+        col_widths=None,
     ):
         """
         Apply a function along both axes.
@@ -411,6 +413,12 @@ class PandasOnRayFramePartitionManager(GenericRayFramePartitionManager):
             List of column partitions.
         item_to_distribute : item, optional
             The item to split up so it can be applied over both axes.
+        row_lengths : list of ints, optional
+            Lengths of partitions for every row. If not specified this information
+            is extracted from partitions itself.
+        col_widths : list of ints, optional
+            Widths of partitions for every column. If not specified this information
+            is extracted from partitions itself.
 
         Returns
         -------
@@ -431,6 +439,8 @@ class PandasOnRayFramePartitionManager(GenericRayFramePartitionManager):
             row_partitions_list,
             col_partitions_list,
             item_to_distribute,
+            row_lengths,
+            col_widths,
         )
 
     @classmethod