… modin/core/storage_formats/base/query_compiler.py
… modin/experimental/core/execution/native/implementations/omnisci_on_native/dataframe/dataframe.py
… modin/pandas/indexing.py
… modin/pandas/test/dataframe/test_indexing.py
… modin/test/storage_formats/pandas/test_internals.py