Signed-off-by: Dmitry Chigarev <dmitry.chigarev@intel.com>

<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/CONTRIBUTING.html
if you have questions about contributing.
-->

## What do these changes do?

Implementation of the idea discussed at the [modin discuss](https://discuss.modin.org/t/does-the-backends-masking-interface-support-slice-usage/227).

This PR aligns masking implementation via `loc/iloc` with the written docstrings. This PR gets rid of passing slices to the backend as indexers, instead of them, it passes `pandas.RangeIndex` which is list-like (doc-string requires this) and has all the benefits of slices at the same time.

- [x] commit message follows format outlined [here](https://modin.readthedocs.io/en/latest/contributing.html)
- [x] passes `flake8 modin`
- [x] passes `black --check modin`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #3519 <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/developer/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->

## Performance changes

With the following ASV benchmarks added:
<details><summary>extra benchmarks</summary>

```python
class TimeIndexingColumns:
    param_names = ["shape"]
    params = [
        get_benchmark_shapes("TimeIndexing")
    ]

    def setup(self, shape):
        self.df = generate_dataframe(ASV_USE_IMPL, "int", *shape, RAND_LOW, RAND_HIGH)
        self.numeric_indexer = [0, 1]
        self.labels_indexer = self.df.columns[self.numeric_indexer].tolist()
    
    def time_iloc(self, shape):
        execute(self.df.iloc[:, self.numeric_indexer])
    
    def time_loc(self, shape):
        execute(self.df.loc[:, self.labels_indexer])
    
    def time___getitem__(self, shape):
        execute(self.df[self.labels_indexer])
```

</details>

Results are the following:

```
      before           after         ratio
     [929c071b]       [711677c9]
     <master>       <range_mask>
         10.4±4ms         12.2±2ms    ~1.17  benchmarks.TimeIndexing.time_loc([1000000, 10], 'scalar')
         20.6±2ms         22.4±2ms     1.09  benchmarks.TimeIndexingColumns.time___getitem__([1000000, 10])
       31.9±0.3ms         34.3±3ms     1.08  benchmarks.TimeIndexing.time_loc([5000, 5000], 'list')
       22.6±0.3ms         24.0±1ms     1.06  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'list')
         17.2±2ms         18.1±1ms     1.05  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'scalar')
          235±3ms          247±3ms     1.05  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'function')
       20.8±0.8ms         21.8±1ms     1.05  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'scalar')
         20.6±2ms         21.0±2ms     1.02  benchmarks.TimeIndexingColumns.time_loc([1000000, 10])
          638±3ms          634±1ms     0.99  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'list')
          782±8ms          778±5ms     0.99  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'function')
       1.11±0.01s       1.10±0.01s     0.99  benchmarks.TimeIndexing.time_loc([1000000, 10], 'list')
          314±6ms          312±8ms     0.99  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool')
       11.9±0.4ms       11.7±0.6ms     0.99  benchmarks.TimeIndexingColumns.time_loc([5000, 5000])
          392±7ms          385±6ms     0.98  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool')
       29.3±0.9ms       28.5±0.5ms     0.97  benchmarks.TimeIndexing.time_loc([5000, 5000], 'scalar')
       12.5±0.5ms       12.2±0.5ms     0.97  benchmarks.TimeIndexingColumns.time___getitem__([5000, 5000])
          459±7ms          444±3ms     0.97  benchmarks.TimeIndexing.time_loc([1000000, 10], 'slice')
          149±6ms          144±9ms     0.97  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool')
          146±6ms          141±2ms     0.97  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool')
         968±20ms         931±20ms     0.96  benchmarks.TimeIndexing.time_loc([1000000, 10], 'function')
          148±7ms          142±7ms     0.96  benchmarks.TimeIndexing.time_loc([5000, 5000], 'slice')
       13.5±0.3ms       12.8±0.4ms     0.95  benchmarks.TimeIndexingColumns.time_iloc([5000, 5000])
         329±10ms          313±5ms     0.95  benchmarks.TimeIndexing.time_loc([5000, 5000], 'function')
          141±8ms          134±8ms     0.95  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'slice')
         23.0±2ms         21.4±1ms     0.93  benchmarks.TimeIndexingColumns.time_iloc([1000000, 10])
-         245±4ms          167±3ms     0.68  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'slice')
```
Overall, performance seems to stay the same.

Extra benchmarks of `apply_func_to_indices_both_axis` with slice indexers, shows that the performance depending on caches hasn't changed:

<details><summary>Ray engine</summary>

|branch|10mln rows, 10 cols (has cache, 112 partitions) | 10mln rows, 10 cols (no cache, 112 partitions) | 1024x1024 (has cache, 1024 partitions) | 1024x1024 (no cache, 1024 partitions) |
|-|-|-|-|-|
|master| 6.23 | 6.21 | 0.10 | 0.08 |
|#3520| 0.63 | 0.78 | 0.09 | 0.06 |

|branch|10mln rows, 10 cols (has cache, 112 partitions) | 10mln rows, 10 cols (has neither frame's nor partition's cache, 112 partitions) | 1024x1024 (has cache, 1024 partitions) | 1024x1024 (has neither frame's nor partition's cache, 1024 partitions) |
|-|-|-|-|-|
|master| 6.07 | 10.53 | 0.10 | 0.43 |
|#3520| 0.60 | 6.22 | 0.10 |  0.32 |

</details>

<details><summary>Dask engine</summary>

|branch|10mln rows, 10 cols (has cache, 112 partitions) | 10mln rows, 10 cols (no cache, 112 partitions) | 1024x1024 (has cache, 1024 partitions) | 1024x1024 (no cache, 1024 partitions) |
|-|-|-|-|-|
|master| 4.71 | 5.04 | 0.23 | 0.24 |
|#3520| 0.85 | 1.07 | 0.24 | 0.24 |

|branch|10mln rows, 10 cols (has cache, 112 partitions) | 10mln rows, 10 cols (has neither frame's nor partition's cache, 112 partitions) | 1024x1024 (has cache, 1024 partitions) | 1024x1024 (has neither frame's nor partition's cache, 1024 partitions) |
|-|-|-|-|-|
|master| 4.64 | 11.21 | 0.24 | 0.84 |
|#3520| 0.85 | 7.34 | 0.25 |  0.82 |

</details>

<details><summary>Reproducer</summary>

```python
import modin.pandas as pd
import numpy as np
from timeit import default_timer as timer

NROWS = 10_000_000
NCOLS = 10

def check_cache(df):
    frame = df._query_compiler._modin_frame
    has_width_cache = 0
    has_length_cache = 0
    for row in frame._partitions:
        for part in row:
            if part._length_cache is not None:
                has_length_cache += 1
            if part._width_cache is not None:
                has_width_cache += 1
    print(f"Total amount of partitions: {frame._partitions.shape[0] * frame._partitions.shape[1]}")
    print(f"\tNumber of partitions having width cache: {has_width_cache}")
    print(f"\tNumber of partitions having length cache: {has_length_cache}")

def check_frame_cache(df):
    frame = df._query_compiler._modin_frame
    print(f"Frame's column widths cache: {frame._column_widths_cache}")
    print(f"Frame's row lengths cache: {frame._row_lengths_cache}")

df = pd.DataFrame({f"col{i}": np.arange(NROWS) for i in range(NCOLS)})

value = df.values

check_cache(df)
check_frame_cache(df)

t1 = timer()
df.iloc[:, :] = value
repr(df)
print(f"has cache: {timer() - t1}")

# Axis relabeling resets partition's cache
df.index = df.index
df._query_compiler._modin_frame._column_widths_cache = None
df._query_compiler._modin_frame._row_lengths_cache = None
check_cache(df)
check_frame_cache(df)

t1 = timer()
df.iloc[:, :] = value
repr(df)
print(f"no cache: {timer() - t1}")

# Possible output:
# Total amount of partitions: 112
#         Number of partitions having width cache: 112
#         Number of partitions having length cache: 112
# Frame's column widths cache: [10]
# Frame's row lengths cache: [89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89286, 89254]
# has cache: 0.6110919890925288
# Total amount of partitions: 112
#         Number of partitions having width cache: 0
#         Number of partitions having length cache: 0
# Frame's column widths cache: None
# Frame's row lengths cache: None
# no cache: 6.532485898584127


```

</details>

You can find more info about why it is required to test in [this thread](https://github.com/modin-project/modin/pull/3520#discussion_r736543884).